feat: create request classes and validate in product

+++ product/src/main/java/com/vr61v/model/Product.java
@@ -20,10 +20,10 @@ public class Product {
     private String name;
 
     @Field(value = "price")
-    private Integer price;
+    private Double price;
 
     @Field(value = "weight")
-    private Integer weight;
+    private Double weight;
 
     @Field(value = "category")
     private String category;
